cuda-requirements.txt
cupy-cuda12x
numba
numpy
opencv-python